requirements.txt
websocket-client==1.2.1
matplotlib==3.4.3
